requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.9.0
networkx>=2.6.0
matplotlib>=3.4.0
tqdm>=4.62.0
pyyaml>=6.0
sentence-transformers>=2.0.0
bertopic>=0.9.0